Update ca_models.py
Add ruleset_changes attribute to keep track of how many times rulesets change for the same grid.

Use different random distribution for initializing color channels with. Random uniform makes for a lot of near-grey colors. Random pareto has a distribution that is mostly close to 0 that will clip to 60,  almost none in the middle, and with very few among very high numbers that will end up getting clipped to 250. This results in the initial grey background image that is heavily glittered with varying tints of intensely red, green, or blue pixels.