Add rule cycling timer feature
Add feature to allow user to set a timer that will alternate between rulesets in a list.

Add handlers for setting the timer and rules to repeatedly add the RULECHANGEEVENT at set intervals to the event queue, and other handlers to handle when the RULECHANGEEVENT event is in the event queue.