Update session_models.py
Add several SessionConfigurationManager methods to get it working with the other new models in this file.

Add methods for making a new default or custom configurations and for starting a new session from the default or a custom configuration settings file.

Renme seed_file_path variable names to seed_image_path in SessionConfiguration and CellularAutomatonSession models.